+++ Plugins/BatchProcessing/src/main/java/de/fraunhofer/iosb/ilt/frostserver/plugin/batchprocessing/batch/Content.java
@@ -59,5 +59,5 @@ public interface Content {
      *
      * @return the headers.
      */
-    public Map<String, String> getHeaders();
+    public Map<String, List<String>> getHeaders();
 }