+++ Plugins/BatchProcessing/src/main/java/de/fraunhofer/iosb/ilt/frostserver/plugin/batchprocessing/multipart/StringContent.java
@@ -86,7 +86,7 @@ public class StringContent implements MultipartContent {
     }
 
     @Override
-    public Map<String, String> getHeaders() {
+    public Map<String, List<String>> getHeaders() {
         return Collections.emptyMap();
     }