+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fixed GeoJSON output not containing geometries.
 * Fixed $select not properly working on unitOfMeasurement sub-properties.
 * Fixed #1934, incorrect URL parameter in docs.
+* Changed header handling to allow for duplicate headers.
 
 
 ## Release version 2.3.0